refactor interp player
