+++ frontend/src/components/interpretation_player/InterpretationPlayer.vue
@@ -1,15 +1,54 @@
 <script setup>
 import { Icon } from '@iconify/vue';
-import { onKeyStroke } from '@vueuse/core';
-import colormap from 'colormap';
-import Peaks from 'peaks.js';
-import { computed, reactive, ref, watch } from 'vue';
+import { onMounted } from 'vue';
 import Popper from 'vue3-popper';
-import { showAlert } from '../../alerts';
-import { api } from '../../axiosInstance';
-import { useAudioStore, useMeasureData, useModulesVisible, useTracksFromDb } from '../../globalStores';
+import { useMeasureData, useModulesVisible, useTracksFromDb } from '../../globalStores';
 import { pinia } from '../../piniaInstance';
-import { getEndMeasure, getSecureConfig, getStartMeasure, truncateFilename } from '../../sharedFunctions';
+import { getEndMeasure, getStartMeasure, getTimeString, truncateFilename } from '../../sharedFunctions';
+
+import {
+    allPeaksReady,
+    differenceRegions,
+    interpretationPlayerOpened,
+    isPlaying,
+    measuresVisible,
+    oneVsRestRelevance,
+    peaksInstancesReady,
+    percLoaded,
+    playing,
+    regionLengths,
+    relevantMeasures,
+    selectedLabel,
+    selectedRegions,
+    selectedRelevanceFeatureName,
+    trackLabels,
+    trackTimes,
+    volume,
+    zoomingEnabled,
+} from './javascript/variables';
+
+import {
+    idxArray,
+    initPlayer,
+    peaksInstances,
+    playPause,
+    resetPlayer,
+    rewind,
+    selectedMeasureData,
+    selectPeaks,
+    toggleMeasures,
+} from './javascript/player';
+
+import { selectRegion } from './javascript/regions';
+
+import {
+    selectDefaultRelevanceFeature,
+    selectRelevanceFeature,
+    selectRelevanceLabel,
+    selectRelevantMeasures,
+} from './javascript/relevance';
+
+import { addControls } from './javascript/controls';
 
 import LoadingWindow from '../LoadingWindow.vue';
 import ModuleTemplate from '../ModuleTemplate.vue';
@@ -21,788 +60,167 @@ import RelevantMeasure from './subcomponents/RelevantMeasure.vue';
 // pinia stores
 const modulesVisible = useModulesVisible(pinia);
 const tracksFromDb = useTracksFromDb(pinia);
-const audioStore = useAudioStore(pinia);
 const measureData = useMeasureData(pinia);
 
-const colors = colormap({
-    colormap: 'viridis',
-    format: 'hex',
-    nshades: 101,
+onMounted(() => {
+    addControls();
 });
 
-const allPeaksReady = computed(() => {
-    return !peaksInstancesReady.value.includes(false);
-});
-
-const numPeaksLoaded = ref(0);
-const percLoaded = ref(0);
-watch(numPeaksLoaded, () => {
-    percLoaded.value = Math.round((numPeaksLoaded.value / (tracksFromDb.syncTracks.length - 1)) * 100);
-});
-
-let peaksInstances = [];
-let selectedMeasureData = [];
-let idxArray = [];
-
-let measureCount = 0;
-let activePeaksIdx = 0;
-let regionSelected = false;
-let interpPlayerOpened = false;
-let syncPoints = null;
-
-let playing = reactive([]);
-
-const isPlaying = ref(false);
-const regionOverlay = ref([]);
-const volume = ref(1);
-
-const currentMeasure = ref(-1);
-const currentRelevance = ref(0);
-const peaksInstancesReady = ref([]);
-const selectedRelevanceData = ref([]);
-const oneVsRestRelevance = ref([]);
-const trackLabels = ref([]);
-const trackTimes = ref([]);
-const regionLengths = ref([]);
-const selectedRelevanceFeature = ref('');
-const regionToSave = ref(false);
-const zoomingEnabled = ref(false);
-
-const audioCtx = new AudioContext();
-const gainNode = audioCtx.createGain();
-gainNode.gain.setValueAtTime(0, audioCtx.currentTime);
-gainNode.connect(audioCtx.destination);
-
-const rampUp = new Float32Array(10);
-const rampDown = new Float32Array(10);
-function createFadeRamps() {
-    for (let i = 0; i < 10; i++) {
-        rampUp[i] = Math.pow(i / 9, 3) * volume.value;
-        rampDown[rampUp.length - i - 1] = Math.pow(i / 9, 3) * volume.value;
-    }
-}
-
-// 10 ms fade in and fade out to prevent clicking when switching recordings
-const fadeIn = () => gainNode.gain.setValueCurveAtTime(rampUp, audioCtx.currentTime, 0.01);
-const fadeOut = () => gainNode.gain.setValueCurveAtTime(rampDown, audioCtx.currentTime, 0.01);
-
-watch(volume, () => {
-    gainNode.gain.setValueAtTime(volume.value, audioCtx.currentTime);
-    createFadeRamps();
-});
-
-// subscribe to pinia modulesVisible state
 modulesVisible.$subscribe((mutation, state) => {
-    if (state.interpretationPlayer && !interpPlayerOpened) {
-        interpPlayerOpened = true;
-        // init helper variables
-        tracksFromDb.syncTracks.forEach((track, idx) => {
-            peaksInstances.push(null);
-            peaksInstancesReady.value.push(false);
-            oneVsRestRelevance.value.push(false);
-            trackLabels.value.push(false);
-            trackTimes.value.push(0);
-            regionLengths.value.push(0);
-            playing.push(false);
-            idxArray.push(idx);
-        });
-
-        audioCtx.resume();
-        createFadeRamps();
-        getMeasureData();
-        getSyncPoints();
-        getAllRegions();
-        selectRelevanceFeature(measureData.relevanceFeatures[0].id);
-
-        setTimeout(() => {
-            tracksFromDb.syncTracks.forEach((track, idx) => {
-                addTrack(track.filename, idx);
-            });
-            addListeners();
-        }, 50);
-        measuresVisible.value = true;
-    } else if (!state.interpretationPlayer && interpPlayerOpened) {
-        if (isPlaying.value) playPause();
-        interpPlayerOpened = false;
-        removeListeners();
-        idxArray = [];
-        peaksInstances = [];
-        selectedMeasureData = [];
-        prevPeaksIdx = null;
-
-        regionOverlay.value = [];
-        peaksInstancesReady.value = [];
-        oneVsRestRelevance.value = [];
-        trackLabels.value = [];
-        trackTimes.value = [];
-        regionLengths.value = [];
-        measuresVisible.value = false;
-        numPeaksLoaded.value = 0;
-        percLoaded.value = 0;
-        selectRelevanceFeature(measureData.relevanceFeatures[0].id);
-
-        playing.fill(false);
+    if (state.interpretationPlayer && !interpretationPlayerOpened.value) {
+        initInterpretationPlayer();
+    } else if (!state.interpretationPlayer && interpretationPlayerOpened.value) {
+        destroyInterpretationPlayer();
     }
 });
 
-onKeyStroke(' ', (e) => {
-    e.preventDefault();
-    if (modulesVisible.interpretationPlayer) playPause();
-});
+function initInterpretationPlayer() {
+    tracksFromDb.syncTracks.forEach((track, idx) => {
+        oneVsRestRelevance.value.push(false);
+        peaksInstancesReady.value.push(false);
+        playing.push(false);
+        regionLengths.value.push(0.0);
+        trackLabels.value.push(false);
+        trackTimes.value.push(0.0);
 
-onKeyStroke('Home', (e) => {
-    if (modulesVisible.interpretationPlayer) rewind();
-});
-
-onKeyStroke(
-    'ArrowUp',
-    (e) => {
-        if (modulesVisible.interpretationPlayer) {
-            const switchIdx = activePeaksIdx - 1;
-            if (switchIdx >= 0) selectPeaks(switchIdx);
-        }
-    },
-    { eventName: 'keyup' }
-);
-
-onKeyStroke(
-    'ArrowDown',
-    (e) => {
-        if (modulesVisible.interpretationPlayer) {
-            const switchIdx = activePeaksIdx + 1;
-            if (switchIdx < peaksInstances.length) selectPeaks(switchIdx);
-        }
-    },
-    { eventName: 'keyup' }
-);
-
-onKeyStroke(
-    'ArrowLeft',
-    (e) => {
-        if (modulesVisible.interpretationPlayer) {
-            const newMeasure = currentMeasure.value - 1;
-            if (newMeasure > -1) goToMeasure(newMeasure);
-        }
-    },
-    { eventName: 'keyup' }
-);
-
-onKeyStroke(
-    'ArrowRight',
-    (e) => {
-        if (modulesVisible.interpretationPlayer) {
-            const newMeasure = currentMeasure.value + 1;
-            if (newMeasure < measureCount) goToMeasure(newMeasure);
-        }
-    },
-    { eventName: 'keyup' }
-);
-
-onKeyStroke(
-    'a',
-    (e) => {
-        if (modulesVisible.interpretationPlayer) {
-            zoomAlign();
-        }
-    },
-    { eventName: 'keyup' }
-);
-
-onKeyStroke('m', (e) => {
-    if (modulesVisible.interpretationPlayer) toggleMeasures();
-});
-
-onKeyStroke('Escape', (e) => {
-    if (modulesVisible.interpretationPlayer) {
-        // if (regionSelected) {
-        //     regionObjects.selected.fill(false);
-        // }
-        regionOverlay.value.fill(false);
-        if (isPlaying.value) playPause();
-        peaksInstances.forEach((peaksInstance, i) => {
-            peaksInstance.segments.removeAll();
-            regionLengths.value[i] = 0;
-            const view = peaksInstance.views.getView('zoomview');
-            view.setZoom({ seconds: 'auto' });
-        });
-        regionToSave.value = false;
-    }
-});
-
-function getMeasureData() {
-    for (let i = 0; i < tracksFromDb.syncTracks.length; i++) {
-        const filename = tracksFromDb.syncTracks[i].filename;
-        if (tracksFromDb.syncTracks[i].gt_measures) {
-            const measures = measureData.getObject(filename).gt_measures;
-            selectedMeasureData.push(measures);
-            for (let j = 1; j < measures.length - 2; j++) {
-                regionOverlay.value.push(false);
-            }
-        } else {
-            selectedMeasureData.push(measureData.getObject(filename).tf_measures);
-        }
-    }
-}
-
-const selectedRegions = ref([]);
-const differenceRegions = ref([]);
-const relevantMeasures = ref([]);
-
-async function getAllRegions() {
-    const selRegionsRes = await api.get('/get-all-regions', getSecureConfig());
-    selectedRegions.value = selRegionsRes.data;
-    const diffRegionsRes = await api.get('/get-diff-regions', getSecureConfig());
-    differenceRegions.value = diffRegionsRes.data.diff_regions;
-}
-
-async function getSyncPoints() {
-    const syncPointsRes = await api.get('/get-sync-points', getSecureConfig());
-    syncPoints = syncPointsRes.data;
-}
-
-function waveformListener(idx, event) {
-    if (idx !== activePeaksIdx) {
-        selectPeaks(idx);
-    }
-}
-
-function addTrack(filename, idx) {
-    const audioElement = document.getElementById(`audio-${idx}`);
-    const audio = audioStore.getAudio(filename);
-    audioElement.src = URL.createObjectURL(audio);
-    const waveformContainer = document.getElementById(`track-div-${idx}`);
-    waveformContainer.addEventListener('mousedown', waveformListener.bind(event, idx));
-    const audioSource = audioCtx.createMediaElementSource(audioElement);
-    audioSource.connect(gainNode);
-    const waveformData = audioStore.getWaveformData(filename);
-    const trackLengthSec = tracksFromDb.getObject(filename).length_sec;
-    // const zoomLevels = createZoomLevels(waveformContainer.offsetWidth, trackLengthSec);
-    const options = {
-        zoomview: {
-            segmentOptions: {
-                style: 'overlay',
-                overlayOffset: 0,
-                overlayOpacity: 0.15,
-                overlayCornerRadius: 0,
-            },
-            container: waveformContainer,
-            playheadColor: 'red',
-            playheadClickTolerance: 500,
-            waveformColor: 'rgb(17 24 39)',
-            axisLabelColor: 'rgb(17 24 39)',
-            axisGridlineColor: 'rgb(17 24 39)',
-            fontFamily: 'Inter',
-        },
-        mediaElement: audioElement,
-        dataUri: {
-            arraybuffer: URL.createObjectURL(waveformData),
-        },
-        showAxisLabels: true,
-        emitCueEvents: true,
-        fontSize: 12,
-        zoomLevels: [1024],
-    };
-    Peaks.init(options, (err, peaks) => {
-        if (err) console.log(err);
-        peaksInstances[idx] = peaks;
-        if (idx === 0) {
-            selectPeaks(idx);
-            selectRelevanceLabel(idx, 'oneVsRest');
-            selectedRelevanceFeatureStr.value = measureData.relevanceFeatures[0].name;
-            measureCount = measureData.relevance.duration.oneVsRest[idx].measureRelevance.length;
-            oneVsRestRelevance.value[0] = true;
-            trackLabels.value[0] = true;
-        }
-        if (filename === tracksFromDb.refTrack.filename) {
-            peaksInstances[idx].on('player.timeupdate', (time) => {
-                trackTimes.value[idx] = time;
-                // add 1 ms to the time to indicate the proper measure
-                const measureIdx = getStartMeasure(time + 0.001);
-                currentMeasure.value = measureIdx - 2;
-                if (currentMeasure.value >= 0) {
-                    currentRelevance.value = selectedRelevanceData.value[currentMeasure.value].relevance.toFixed(2);
-                }
-            });
-        } else {
-            peaksInstances[idx].on('player.timeupdate', (time) => {
-                trackTimes.value[idx] = time;
-            });
-        }
-        addMeasuresToPeaksInstance(idx);
-        peaksInstancesReady.value[idx] = true;
-        numPeaksLoaded.value += 1;
-        const view = peaksInstances[idx].views.getView('zoomview');
-        view.setZoom({ seconds: trackLengthSec });
+        // from player
+        idxArray.push(idx);
+        peaksInstances.push(null);
     });
+
+    initPlayer();
+    selectDefaultRelevanceFeature();
+    measuresVisible.value = true;
+    interpretationPlayerOpened.value = true;
 }
 
-async function selectRegion(regionIdx, obj) {
-    hideAllRegions();
-    regionToSave.value = false;
-    // regionObjects.selected[regionIdx] = !regionObjects.selected[regionIdx];
-    // regionObjects.selected.forEach((arrayValue, i) => {
-    //     if (i !== regionIdx) regionObjects.selected[i] = false;
-    // });
-    const referenceName = tracksFromDb.refTrack.filename;
-    const refIdx = tracksFromDb.getIdx(referenceName);
-    fadeOut();
-    await sleep(10);
-    peaksInstances[activePeaksIdx].player.pause();
-    isPlaying.value = false;
-    regionOverlay.value.fill(false);
-    // if (regionObjects.selected[regionIdx]) {
-    const startIdx = findClosestTimeIdx(refIdx, obj.startTime);
-    const endIdx = findClosestTimeIdx(refIdx, obj.endTime);
-    switch (obj.type) {
-        case 'selectedRegion':
-            addSelectedRegion(startIdx, endIdx, obj);
-            break;
-        case 'differenceRegion':
-            const targetIdx = tracksFromDb.getIdx(obj.regionName);
-            addDifferenceRegion(refIdx, targetIdx, obj);
-            break;
-        case 'relevantMeasure':
-            addRelevantMeasure(obj);
-            break;
-    }
-    // } else {
-    //     hideAllRegions();
-    //     isPlaying.value = false;
-    // }
+function destroyInterpretationPlayer() {
+    // variables
+    oneVsRestRelevance.value = [];
+    peaksInstancesReady.value = [];
+    playing.value = [];
+    regionLengths.value = [];
+    trackLabels.value = [];
+    trackTimes.value = [];
+
+    // player
+    idxArray.splice(0);
+    peaksInstances.splice(0);
+    selectedMeasureData.splice(0);
+
+    resetPlayer();
+    selectDefaultRelevanceFeature();
+    measuresVisible.value = false;
+    interpretationPlayerOpened.value = false;
 }
 
-function addSelectedRegion(startIdx, endIdx, obj) {
-    for (let i = 0; i < peaksInstances.length; i++) {
-        peaksInstances[i].segments.add({
-            color: obj.color,
-            borderColor: obj.color,
-            startTime: syncPoints[i][startIdx],
-            endTime: syncPoints[i][endIdx],
-            id: 'selectedRegion',
-        });
-    }
-    const startMeasure = getStartMeasure(obj.startTime) - 1;
-    const endMeasure = getEndMeasure(obj.endTime) - 1;
-    for (let j = startMeasure; j < endMeasure; j++) {
-        regionOverlay.value[j] = true;
-    }
-    zoomOnSelectedRegion();
-}
+// const measureMessage = ref(null);
+// function logMeasure(i) {
+//     measureMessage.value =
+//         'Measure: ' + (i + 1) + ', Relevance: ' + selectedRelevanceData.value[i].relevance.toFixed(2);
+//     const popup = document.getElementById('measure-popup');
+//     const width = measureCount * 16;
+//     const topBar = document.getElementById('top-bar');
+//     popup.style.left =
+//         Math.round(((i + 1) / measureCount) * width) - topBar.scrollLeft - popup.offsetWidth / 2 + 18 + 'px';
+// }
 
-function addRelevantMeasure(obj) {
-    for (let i = 0; i < peaksInstances.length; i++) {
-        peaksInstances[i].segments.add({
-            color: obj.color,
-            borderColor: obj.color,
-            startTime: selectedMeasureData[i][obj.measureIdx],
-            endTime: selectedMeasureData[i][obj.measureIdx + 1],
-            id: 'relevantMeasure',
-        });
-    }
-    regionOverlay.value[obj.measureIdx] = true;
-    zoomOnMeasureSelection(obj.measureIdx, obj.measureIdx);
-}
+// function clearMessage() {
+//     measureMessage.value = null;
+// }
 
-function addDifferenceRegion(refIdx, targetIdx, obj) {
-    selectPeaks(refIdx);
-    peaksInstances[refIdx].player.seek(obj.startTime);
-    peaksInstances[refIdx].segments.add({
-        color: obj.color,
-        borderColor: obj.color,
-        startTime: obj.startTime,
-        endTime: obj.endTime,
-        id: 'selectedRegion',
-    });
-    peaksInstances[targetIdx].segments.add({
-        color: obj.color,
-        borderColor: obj.color,
-        startTime: obj.startTimeTarget,
-        endTime: obj.endTimeTarget,
-        id: 'selectedRegion',
-    });
-    zoomOnSelectedRegion();
-}
+// let firstMeasure = null;
+// let isHoldingMouseButton = false;
+// let dragged = false;
+// function addListeners() {
+//     const relevanceBar = document.getElementById('overview-2');
+//     relevanceBar.addEventListener('mousedown', relevanceBarMouseDown);
+//     window.addEventListener('mouseup', relevanceBarMouseUp);
+//     window.addEventListener('mousemove', relevanceBarMouseMove);
+//     const scrollContainer = document.getElementById('top-bar');
+//     scrollContainer.addEventListener('wheel', horizontalScroll);
+//     const container = document.getElementById('audio-container');
+//     // container.addEventListener('wheel', (e) => {
+//     //     e.preventDefault();
+//     // });
+//     window.addEventListener('keydown', (e) => {
+//         if (e.key === 'ArrowUp' || e.key === 'ArrowDown') {
+//             e.preventDefault();
+//         }
+//     });
+//     // container.addEventListener('mousewheel', (event) => {
+//     //     if (event.deltaY < 0) {
+//     //         peaksInstances.forEach((peaksInstance) => {
+//     //             peaksInstance.zoom.zoomIn();
+//     //             zoomAlign();
+//     //         });
+//     //     } else {
+//     //         peaksInstances.forEach((peaksInstance) => {
+//     //             peaksInstance.zoom.zoomOut();
+//     //             zoomAlign();
+//     //         });
+//     //     }
+//     // });
+// }
 
-function hideAllRegions() {
-    peaksInstances.forEach((peaksInstance) => {
-        peaksInstance.segments.removeAll();
-        regionLengths.value.fill(0);
-    });
-    regionSelected = false;
-}
+// function removeListeners() {
+//     const relevanceBar = document.getElementById('overview-2');
+//     relevanceBar.removeEventListener('mousedown', relevanceBarMouseDown);
+//     window.removeEventListener('mouseup', relevanceBarMouseUp);
+//     window.removeEventListener('mousemove', relevanceBarMouseMove);
+//     const scrollContainer = document.getElementById('top-bar');
+//     scrollContainer.removeEventListener('wheel', horizontalScroll);
+// }
 
-function findClosestTimeIdx(peaksIdx, time) {
-    const closestTime = syncPoints[peaksIdx].reduce((prev, curr) =>
-        Math.abs(curr - time) < Math.abs(prev - time) ? curr : prev
-    );
-    return syncPoints[peaksIdx].indexOf(closestTime);
-}
+// function relevanceBarMouseDown(event) {
+//     const relevanceBar = document.getElementById('overview-2');
+//     const bounds = relevanceBar.getBoundingClientRect();
+//     const barWidth = 16 * measureCount;
+//     event.preventDefault();
+//     isHoldingMouseButton = true;
+//     const x = event.clientX - bounds.left;
+//     const dragOverMeasureIdx = Math.round((x / barWidth) * measureCount);
+//     firstMeasure = dragOverMeasureIdx;
+// }
 
-function seekCallback(time) {
-    const closestTimeIdx = findClosestTimeIdx(activePeaksIdx, time);
-    selectedIndices.forEach((idx) => {
-        peaksInstances[idx].player.seek(syncPoints[idx][closestTimeIdx]);
-    });
-    // if (time >= selectedMeasureData[activePeaksIdx][repeatMeasureIdxEnd.value] - 0.05 && regionSelected) {
-    //     goToMeasure(repeatMeasureIdxStart.value);
-    // }
-}
+// function relevanceBarMouseUp() {
+//     isHoldingMouseButton = false;
+//     const startMeasureIdx = regionOverlay.value.indexOf(true);
+//     const endMeasureIdx = regionOverlay.value.lastIndexOf(true);
+//     if (dragged) {
+//         regionToSave.value = true;
+//         const referenceName = tracksFromDb.refTrack.filename;
+//         const refIdx = tracksFromDb.getIdx(referenceName);
+//         zoomOnMeasureSelection(startMeasureIdx, endMeasureIdx);
+//         repeatMeasureIdxStart.value = startMeasureIdx;
+//         repeatMeasureIdxEnd.value = endMeasureIdx + 2;
+//         startTime.value = selectedMeasureData[refIdx][startMeasureIdx + 1];
+//         endTime.value = selectedMeasureData[refIdx][endMeasureIdx + 2];
+//     }
+//     dragged = false;
+// }
 
-let prevPeaksIdx = null;
-let selectedIndices = null;
-async function selectPeaks(idx) {
-    selectedIndices = idxArray.slice();
-    selectedIndices.splice(idx, 1);
-    if (prevPeaksIdx !== null) {
-        playing[prevPeaksIdx] = false;
-        if (isPlaying.value) {
-            fadeOut();
-            await sleep(10);
-            peaksInstances[prevPeaksIdx].player.pause();
-        }
-        peaksInstances[prevPeaksIdx].off('player.timeupdate', seekCallback);
-    }
-    // add seekCallback to peaks instance specified by idx
-    playing[idx] = true;
-    activePeaksIdx = idx;
-    peaksInstances[idx].on('player.timeupdate', seekCallback);
-    // if playing is active
-    if (isPlaying.value) {
-        // play currently selected region if it is not null
-        const selectedRegion = peaksInstances[idx].segments.getSegment('selectedRegion');
-        if (selectedRegion !== null) {
-            const closestTimeIdx = findClosestTimeIdx(prevPeaksIdx, trackTimes.value[prevPeaksIdx]);
-            peaksInstances[idx].player.playSegment(selectedRegion, true);
-            peaksInstances[idx].player.seek(syncPoints[idx][closestTimeIdx]);
-            await sleep(10);
-            fadeIn();
-        }
-        // otherwise just continue playing at the current cursor position
-        else {
-            peaksInstances[idx].player.play();
-            await sleep(10);
-            fadeIn();
-        }
-    }
-    prevPeaksIdx = idx;
-}
+// function relevanceBarMouseMove(event) {
+//     const relevanceBar = document.getElementById('overview-2');
+//     const bounds = relevanceBar.getBoundingClientRect();
+//     const barWidth = 16 * measureCount;
+//     if (isHoldingMouseButton && !regionSelected) {
+//         dragged = true;
+//         regionOverlay.value.fill(false);
+//         const x = event.clientX - bounds.left;
+//         let dragOverMeasureIdx = Math.floor((x / barWidth) * measureCount);
+//         if (dragOverMeasureIdx > measureCount) dragOverMeasureIdx = measureCount - 1;
+//         for (let i = firstMeasure; i < dragOverMeasureIdx + 1; i++) {
+//             regionOverlay.value[i] = true;
+//         }
+//     }
+// }
 
-function sleep(ms) {
-    return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-async function playPause() {
-    // pause playing if it is active
-    if (isPlaying.value) {
-        fadeOut();
-        await sleep(10);
-        peaksInstances[activePeaksIdx].player.pause();
-    } else {
-        // play currently selected region if it is not null
-        const selectedRegion = peaksInstances[activePeaksIdx].segments.getSegment('selectedRegion');
-        if (selectedRegion !== null) {
-            peaksInstances[activePeaksIdx].player.playSegment(selectedRegion, true);
-            fadeIn();
-        }
-        // otherwise just continue playing at the current cursor position
-        else {
-            peaksInstances[activePeaksIdx].player.play();
-            fadeIn();
-        }
-    }
-    isPlaying.value = !isPlaying.value;
-}
-
-async function rewind() {
-    fadeOut();
-    await sleep(20);
-    peaksInstances[activePeaksIdx].player.seek(0);
-    fadeIn();
-}
-
-function zoomAlign() {
-    for (let i = 0; i < peaksInstances.length; i++) {
-        if (peaksInstances[i].segments.getSegment('selectedRegion') === null) {
-            continue;
-        }
-        const view = peaksInstances[i].views.getView('zoomview');
-        const segment = peaksInstances[i].segments.getSegment('selectedRegion');
-        view.setStartTime(segment.startTime);
-    }
-}
-
-function getLongestRegion() {
-    let secs = -1;
-    for (let i = 0; i < peaksInstances.length; i++) {
-        const segment = peaksInstances[i].segments.getSegment('selectedRegion');
-        const seglen = segment.endTime - segment.startTime;
-        if (seglen > secs) secs = seglen + 1;
-    }
-    return secs;
-}
-
-function zoomOnSelectedRegion() {
-    const segment = peaksInstances[activePeaksIdx].segments.getSegment('selectedRegion');
-    const secs = getLongestRegion();
-    peaksInstances[activePeaksIdx].player.seek(segment.startTime);
-    for (let i = 0; i < peaksInstances.length; i++) {
-        const view = peaksInstances[i].views.getView('zoomview');
-        const segment = peaksInstances[i].segments.getSegment('selectedRegion');
-        regionLengths.value[i] = segment.endTime - segment.startTime;
-        if (zoomingEnabled.value) {
-            view.setZoom({ seconds: secs });
-            view.setStartTime(segment.startTime);
-        }
-    }
-}
-
-async function zoomOnMeasureSelection(startMeasureIdx, endMeasureIdx) {
-    if (isPlaying.value) await playPause();
-    regionSelected = true;
-    hideAllRegions();
-    peaksInstances[activePeaksIdx].player.seek(selectedMeasureData[activePeaksIdx][startMeasureIdx + 1]);
-    for (let i = 0; i < peaksInstances.length; i++) {
-        const view = peaksInstances[i].views.getView('zoomview');
-        view.enableAutoScroll(false, {});
-        peaksInstances[i].segments.add({
-            color: 'blue',
-            borderColor: 'blue',
-            startTime: selectedMeasureData[i][startMeasureIdx + 1],
-            endTime: selectedMeasureData[i][endMeasureIdx + 2],
-            id: 'selectedRegion',
-        });
-        regionLengths.value[i] =
-            selectedMeasureData[i][endMeasureIdx + 2] - selectedMeasureData[i][startMeasureIdx + 1];
-    }
-    const secs = getLongestRegion();
-    for (let i = 0; i < peaksInstances.length; i++) {
-        const view = peaksInstances[i].views.getView('zoomview');
-        if (zoomingEnabled.value) {
-            view.setZoom({ seconds: secs });
-            view.setStartTime(selectedMeasureData[i][startMeasureIdx + 1]);
-        }
-    }
-}
-
-function getTimeString(seconds, start, end) {
-    return new Date(seconds * 1000).toISOString().slice(start, end);
-}
-
-function selectRelevantMeasures() {
-    let sortedRelevance = selectedRelevanceData.value.slice(0);
-    sortedRelevance.sort((a, b) => a.relevance - b.relevance).reverse();
-    relevantMeasures.value = sortedRelevance.slice(0, 10);
-}
-
-const measuresVisible = ref(false);
-function toggleMeasures() {
-    if (measuresVisible.value) {
-        peaksInstances.forEach((peaksInstance) => {
-            peaksInstance.points.removeAll();
-        });
-        measuresVisible.value = false;
-    } else {
-        for (let i = 0; i < tracksFromDb.syncTracks.length; i++) {
-            addMeasuresToPeaksInstance(i);
-        }
-        measuresVisible.value = true;
-    }
-}
-
-function addMeasuresToPeaksInstance(idx) {
-    for (let j = 1; j < selectedMeasureData[idx].length - 1; j++) {
-        let labelText = `${j}`;
-        peaksInstances[idx].points.add({
-            time: selectedMeasureData[idx][j],
-            labelText: labelText,
-            editable: false,
-            color: 'rgb(0, 0, 200)',
-        });
-    }
-}
-
-let canRewind = true;
-async function goToMeasure(measureIdx) {
-    if (canRewind) {
-        canRewind = false;
-        fadeOut();
-        await sleep(20);
-        const seekTime = selectedMeasureData[activePeaksIdx][measureIdx + 1];
-        peaksInstances[activePeaksIdx].player.seek(seekTime);
-        fadeIn();
-        canRewind = true;
-    }
-}
-
-const measureMessage = ref(null);
-function logMeasure(i) {
-    measureMessage.value =
-        'Measure: ' + (i + 1) + ', Relevance: ' + selectedRelevanceData.value[i].relevance.toFixed(2);
-    const popup = document.getElementById('measure-popup');
-    const width = measureCount * 16;
-    const topBar = document.getElementById('top-bar');
-    popup.style.left =
-        Math.round(((i + 1) / measureCount) * width) - topBar.scrollLeft - popup.offsetWidth / 2 + 18 + 'px';
-}
-
-function clearMessage() {
-    measureMessage.value = null;
-}
-
-let firstMeasure = null;
-let isHoldingMouseButton = false;
-let dragged = false;
-function addListeners() {
-    const relevanceBar = document.getElementById('overview-2');
-    relevanceBar.addEventListener('mousedown', relevanceBarMouseDown);
-    window.addEventListener('mouseup', relevanceBarMouseUp);
-    window.addEventListener('mousemove', relevanceBarMouseMove);
-    const scrollContainer = document.getElementById('top-bar');
-    scrollContainer.addEventListener('wheel', horizontalScroll);
-    const container = document.getElementById('audio-container');
-    // container.addEventListener('wheel', (e) => {
-    //     e.preventDefault();
-    // });
-    window.addEventListener('keydown', (e) => {
-        if (e.key === 'ArrowUp' || e.key === 'ArrowDown') {
-            e.preventDefault();
-        }
-    });
-    // container.addEventListener('mousewheel', (event) => {
-    //     if (event.deltaY < 0) {
-    //         peaksInstances.forEach((peaksInstance) => {
-    //             peaksInstance.zoom.zoomIn();
-    //             zoomAlign();
-    //         });
-    //     } else {
-    //         peaksInstances.forEach((peaksInstance) => {
-    //             peaksInstance.zoom.zoomOut();
-    //             zoomAlign();
-    //         });
-    //     }
-    // });
-}
-
-function removeListeners() {
-    const relevanceBar = document.getElementById('overview-2');
-    relevanceBar.removeEventListener('mousedown', relevanceBarMouseDown);
-    window.removeEventListener('mouseup', relevanceBarMouseUp);
-    window.removeEventListener('mousemove', relevanceBarMouseMove);
-    const scrollContainer = document.getElementById('top-bar');
-    scrollContainer.removeEventListener('wheel', horizontalScroll);
-}
-
-function relevanceBarMouseDown(event) {
-    const relevanceBar = document.getElementById('overview-2');
-    const bounds = relevanceBar.getBoundingClientRect();
-    const barWidth = 16 * measureCount;
-    event.preventDefault();
-    isHoldingMouseButton = true;
-    const x = event.clientX - bounds.left;
-    const dragOverMeasureIdx = Math.round((x / barWidth) * measureCount);
-    firstMeasure = dragOverMeasureIdx;
-}
-
-const repeatMeasureIdxStart = ref(0);
-const repeatMeasureIdxEnd = ref(0);
-function relevanceBarMouseUp() {
-    isHoldingMouseButton = false;
-    const startMeasureIdx = regionOverlay.value.indexOf(true);
-    const endMeasureIdx = regionOverlay.value.lastIndexOf(true);
-    if (dragged) {
-        regionToSave.value = true;
-        const referenceName = tracksFromDb.refTrack.filename;
-        const refIdx = tracksFromDb.getIdx(referenceName);
-        zoomOnMeasureSelection(startMeasureIdx, endMeasureIdx);
-        repeatMeasureIdxStart.value = startMeasureIdx;
-        repeatMeasureIdxEnd.value = endMeasureIdx + 2;
-        startTime.value = selectedMeasureData[refIdx][startMeasureIdx + 1];
-        endTime.value = selectedMeasureData[refIdx][endMeasureIdx + 2];
-    }
-    dragged = false;
-}
-
-function relevanceBarMouseMove(event) {
-    const relevanceBar = document.getElementById('overview-2');
-    const bounds = relevanceBar.getBoundingClientRect();
-    const barWidth = 16 * measureCount;
-    if (isHoldingMouseButton && !regionSelected) {
-        dragged = true;
-        regionOverlay.value.fill(false);
-        const x = event.clientX - bounds.left;
-        let dragOverMeasureIdx = Math.floor((x / barWidth) * measureCount);
-        if (dragOverMeasureIdx > measureCount) dragOverMeasureIdx = measureCount - 1;
-        for (let i = firstMeasure; i < dragOverMeasureIdx + 1; i++) {
-            regionOverlay.value[i] = true;
-        }
-    }
-}
-
-function horizontalScroll(event) {
-    event.preventDefault();
-    const scrollContainer = document.getElementById('top-bar');
-    scrollContainer.scrollLeft += event.deltaY;
-}
-
-const selectedRelevanceFeatureStr = ref('');
-const selectedLabel = ref('');
-const selectedType = ref('');
-const selectedIdx = ref(0);
-
-function selectRelevanceFeature(id, name) {
-    selectedRelevanceFeatureStr.value = name;
-    selectedRelevanceFeature.value = id;
-    if (selectedLabel.value !== '') {
-        selectedRelevanceData.value =
-            measureData.relevance[selectedRelevanceFeature.value][selectedType.value][
-                selectedIdx.value
-            ].measureRelevance.slice();
-        selectedLabel.value =
-            measureData.relevance[selectedRelevanceFeature.value][selectedType.value][selectedIdx.value].labelName;
-    }
-}
-
-function selectRelevanceLabel(idx, type) {
-    oneVsRestRelevance.value.fill(false);
-    trackLabels.value.fill(false);
-    selectedType.value = type;
-    selectedIdx.value = idx;
-    if (type === 'oneVsRest') {
-        oneVsRestRelevance.value[idx] = true;
-        trackLabels.value[idx] = true;
-    } else {
-        trackLabels.value = measureData.relevance[selectedRelevanceFeature.value][type][idx].labels.slice();
-    }
-    selectedRelevanceData.value =
-        measureData.relevance[selectedRelevanceFeature.value][type][idx].measureRelevance.slice();
-    selectedLabel.value = measureData.relevance[selectedRelevanceFeature.value][type][idx].labelName;
-}
-
-const regionName = ref('');
-const startTime = ref(0);
-const endTime = ref(0);
-const beatsPerMeasure = ref(1);
-
-function saveRegion() {
-    if (regionName.value !== '') {
-        const data = {
-            startTime: startTime.value,
-            endTime: endTime.value,
-            regionName: regionName.value,
-            lengthSec: endTime.value - startTime.value,
-            beatsPerMeasure: beatsPerMeasure.value,
-        };
-        api.post(`/save-region`, data, getSecureConfig()).then(() => {
-            regionToSave.value = false;
-            regionName.value = '';
-            beatsPerMeasure.value = 1;
-            showAlert(`Region ${data.regionName} successfully added.`, 1500);
-            getAllRegions();
-        });
-    } else {
-        showAlert('Region must have a name!', 1500);
-    }
-}
+// function horizontalScroll(event) {
+//     event.preventDefault();
+//     const scrollContainer = document.getElementById('top-bar');
+//     scrollContainer.scrollLeft += event.deltaY;
+// }
 </script>
 
 <template>
@@ -819,7 +237,7 @@ function saveRegion() {
         </template>
 
         <template v-slot:module-content>
-            <div class="flex h-[4rem] w-full justify-center border-b dark:border-gray-700">
+            <!-- <div class="flex h-[4rem] w-full justify-center border-b dark:border-gray-700">
                 <div class="w-[calc(100%-3rem)] overflow-y-hidden overflow-x-scroll" id="top-bar">
                     <div
                         id="overview-3"
@@ -871,7 +289,7 @@ function saveRegion() {
                         </div>
                     </div>
                 </div>
-            </div>
+            </div> -->
 
             <div
                 id="settings-bar"
@@ -898,11 +316,8 @@ function saveRegion() {
                             v-for="(obj, i) in selectedRegions"
                             :region-name="obj.regionName"
                             :idx="i"
-                            :start-time="getTimeString(obj.startTime, 14, 22)"
-                            :end-time="getTimeString(obj.endTime, 14, 22)"
                             :start-measure="getStartMeasure(obj.startTime)"
                             :end-measure="getEndMeasure(obj.endTime)"
-                            :beats-per-measure="obj.beatsPerMeasure"
                             @click="selectRegion(i, obj)">
                         </SelectedRegion>
                     </SubMenu>
@@ -929,7 +344,7 @@ function saveRegion() {
                         </RelevantMeasure>
                     </SubMenu>
                 </div>
-                <div
+                <!-- <div
                     v-show="regionToSave"
                     class="flex h-10 flex-row items-center justify-center gap-1 rounded-md bg-neutral-200 px-2 text-sm dark:bg-gray-400 dark:text-black">
                     <input
@@ -956,7 +371,7 @@ function saveRegion() {
                             class="input-field-nomargin h-7 w-12" />
                     </div>
                     <button class="btn btn-blue" @click="saveRegion()">Save region</button>
-                </div>
+                </div> -->
             </div>
 
             <div
@@ -968,7 +383,7 @@ function saveRegion() {
                     <div class="flex flex-row gap-1 text-xs">
                         <p>Selected feature:</p>
                         <p class="flex items-center rounded-md bg-neutral-200 px-2 dark:bg-gray-400 dark:text-black">
-                            {{ selectedRelevanceFeatureStr }}
+                            {{ selectedRelevanceFeatureName }}
                         </p>
                     </div>
