+++ frontend/src/components/shared_components/SelectedRegion.vue
@@ -2,11 +2,8 @@
 defineProps({
     idx: Number,
     regionName: String,
-    startTime: String,
-    endTime: String,
     startMeasure: Number,
     endMeasure: Number,
-    beatsPerMeasure: Number,
 });
 </script>
 
@@ -19,7 +16,7 @@ defineProps({
             {{ regionName }}
         </p>
         <div class="flex h-full gap-2 rounded-md py-1 dark:bg-gray-400 dark:hover:bg-gray-700">
-            <div class="flex flex-row gap-2">
+            <!-- <div class="flex flex-row gap-2">
                 <p class="flex h-full items-center">Reference:</p>
                 <p class="flex w-20 select-none items-center justify-center rounded-md bg-green-500 text-xs text-white">
                     {{ startTime }}
@@ -27,14 +24,10 @@ defineProps({
                 <p class="flex w-20 select-none items-center justify-center rounded-md bg-red-500 text-xs text-white">
                     {{ endTime }}
                 </p>
-            </div>
+            </div> -->
             <p class="flex w-28 select-none items-center justify-center rounded-md bg-neutral-700 text-xs text-white">
-                Measures:
+                Measures
                 {{ startMeasure }}–{{ endMeasure }}
-            </p>
-            <p class="flex w-36 select-none items-center justify-center rounded-md bg-neutral-700 text-xs text-white">
-                Beats per measure:
-                {{ beatsPerMeasure }}
             </p>
         </div>
     </div>
