+++ frontend/src/components/track_manager/subcomponents/BottomLegend.vue
@@ -9,7 +9,7 @@ defineProps({
         class="flex h-[1.75rem] w-full select-none items-center justify-between gap-3 border-b px-7 text-sm dark:border-gray-700">
         <div class="flex items-center justify-start gap-3">
             <div class="flex flex-row items-center justify-center gap-1">
-                <div class="h-2 w-2 rounded-full bg-violet-800"></div>
+                <div class="h-2 w-2 rounded-full bg-violet-400"></div>
                 <p>Reference track</p>
             </div>
             <div class="flex flex-row items-center justify-center gap-1">
